add icons, support indent, rename toc

+++ renderer.js
@@ -6,16 +6,19 @@ let editor = null;
 function init() {
     document.getElementById("book_file").onchange = readbook;
     document.getElementById("book_save").onclick = savebook;
+    document.getElementById("button_open").onclick = () => document.getElementById("book_file").click();
 }
 
 function readbook() {
     let file = document.getElementById("book_file").files[0];
-    if (editor) {
-        editor.clear();
-        editor = null;
+    if (file) {
+        if (editor) {
+            editor.clear();
+            editor = null;
+        }
+        editor = new BookEditor(file.path);
+        document.getElementById("message").innerHTML = "";
     }
-    editor = new BookEditor(file.path);
-    document.getElementById("message").innerHTML = "";
 }
 
 function savebook() {
